Update README.md
a fantasy adventure film that follows Perseus, the mortal son of Zeus, as he embarks on a perilous journey to defeat the monstrous Kraken and save the city of Argos. Set in the mythological world of ancient Greece, the movie features epic battles, fearsome creatures, and the interference of gods and goddesses. Alongside a group of loyal warriors, Perseus faces numerous challenges, including Medusa and other mythical beings, to fulfill his destiny and protect humanity from divine wrath.